Move queries subpackage from core to registry.
No real code changes here; just reorganization and import adjustments.
