+++ python/lsst/daf/butler/registry/__init__.py
@@ -24,3 +24,4 @@ from ._registry import *
 from ._dbAuth import *
 
 from . import interfaces
+from . import queries
