+++ python/lsst/daf/butler/registry/queries/exprParser/ply/README.md
