+++ python/lsst/daf/butler/registry/queries/exprParser/ply/lex.py
